Fix scripts and add documentation
