+++ simulation/functions/functions.list
@@ -1,9 +1,4 @@
 fibonacci-go
 fibonacci-nodejs
 fibonacci-python
-aes-go
-aes-nodejs
-aes-python
-auth-go
-auth-nodejs
-auth-python
+
